+++ bundle/src/main/java/com/adobe/acs/commons/contentsync/ContentReader.java
@@ -19,6 +19,7 @@
  */
 package com.adobe.acs.commons.contentsync;
 
+import org.apache.jackrabbit.util.ISO8601;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -28,25 +29,24 @@ import javax.jcr.Workspace;
 import javax.jcr.nodetype.NodeType;
 import javax.jcr.nodetype.NodeTypeManager;
 import javax.jcr.nodetype.PropertyDefinition;
-import javax.json.Json;
-import javax.json.JsonArray;
-import javax.json.JsonObject;
-import javax.json.JsonObjectBuilder;
-import javax.json.JsonString;
-import javax.json.JsonValue;
+import javax.json.*;
 import java.lang.invoke.MethodHandles;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.*;
+import java.util.regex.Pattern;
 
 import static org.apache.jackrabbit.JcrConstants.JCR_MIXINTYPES;
 import static org.apache.jackrabbit.JcrConstants.JCR_PRIMARYTYPE;
 
 public class ContentReader {
     private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+    private static final DateTimeFormatter ECMA_DATE_FORMAT = DateTimeFormatter.ofPattern("EEE MMM dd yyyy HH:mm:ss 'GMT'Z");
+
+    private static final Pattern ECMA_REGEX = Pattern.compile(
+            "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun) (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec) (\\d{2}) \\d{4} ([01]\\d|2[0-3]):([0-5]\\d):([0-5]\\d) GMT[+-]\\d{4}");
 
     static final String BINARY_DATA_PLACEHOLDER = "0";
 
@@ -102,8 +102,18 @@ public class ContentReader {
                     break;
                 case ARRAY:
                     JsonArray array = (JsonArray) value;
+                    if(name.equals(JCR_MIXINTYPES)){
+                        array = sanitizeMixins((JsonArray) value);
+                    }
                     out.add(name, array);
                     break;
+                case STRING:
+                    String str = ((JsonString)value).getString();
+                    if(ECMA_REGEX.matcher(str).matches()){
+                        // convert to legacy ECMA dates to ISO8601
+                        String isoDate = toISO8601(str);
+                        value = Json.createValue(isoDate);
+                    }
                 default:
                     if (colonIdx == 0) {
                         // Leading colon in Sling GET Servlet JSON designates binary data, e.g. :jcr:data
@@ -151,7 +161,10 @@ public class ContentReader {
         JsonArray mixins = node.getJsonArray(JCR_MIXINTYPES);
         if (mixins != null) {
             for (JsonValue item : mixins) {
-                checkTypes.add(((JsonString) item).getString());
+                String mixin = ((JsonString) item).getString();
+                if(nodeTypeManager.hasNodeType(mixin)){
+                    checkTypes.add(mixin);
+                }
             }
         }
         for (String typeName : checkTypes) {
@@ -164,6 +177,18 @@ public class ContentReader {
         }
 
         return props;
+    }
+
+    JsonArray sanitizeMixins(JsonArray remoteMixins) throws RepositoryException {
+        JsonArrayBuilder mixinTypes = Json.createArrayBuilder();
+        for(JsonValue v : remoteMixins){
+            if(nodeTypeManager.hasNodeType(((JsonString)v).getString())){
+                mixinTypes.add(v);
+            } else {
+                log.debug("skipping unknown mixin: {}", v);
+            }
+        }
+        return mixinTypes.build();
     }
 
     private void collectBinaryProperties(JsonObject node, String parent, List<String> binaryProperties) {
@@ -210,5 +235,23 @@ public class ContentReader {
         List<String> binaryProperties = new ArrayList<>();
         collectBinaryProperties(node, "", binaryProperties);
         return binaryProperties;
+    }
+
+    public static boolean isECMADate(String str) {
+        return ECMA_REGEX.matcher(str).matches();
+    }
+
+    public static Calendar parseEcmaDate(String ecmaDate){
+        try {
+            ZonedDateTime zonedDateTime = ZonedDateTime.parse(ecmaDate, ECMA_DATE_FORMAT);
+            return GregorianCalendar.from(zonedDateTime);
+        } catch(Exception e){
+            return null;
+        }
+    }
+
+    public static String toISO8601(String ecmaDate){
+        Calendar calendar = parseEcmaDate(ecmaDate);
+        return calendar == null ? ecmaDate : ISO8601.format(calendar);
     }
 }