+++ bundle/src/main/java/com/adobe/acs/commons/contentsync/servlet/ContentCatalogServlet.java
@@ -30,20 +30,16 @@ import org.apache.sling.event.jobs.JobManager;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.json.JsonObjectBuilder;
-import javax.json.JsonReader;
-import javax.json.JsonWriter;
+import javax.json.*;
 import javax.servlet.Servlet;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
+import java.io.*;
 import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
 import static com.adobe.acs.commons.contentsync.ContentCatalogJobConsumer.JOB_TOPIC;
+import static com.adobe.acs.commons.contentsync.ContentSyncService.JOB_RESULTS_BASE_PATH;
 
 /**
  * Submits a catalog job or retrieves job status and results.
@@ -107,28 +103,37 @@ public class ContentCatalogServlet extends SlingSafeMethodsServlet {
         response.setContentType("application/json");
 
         JsonObjectBuilder result = Json.createObjectBuilder();
-        String jobId = request.getParameter(JOB_ID);
-        if (jobId == null) {
-            Job job = submitJob(request);
-            result.add(JOB_ID, job.getId());
-            result.add(JOB_STATUS, job.getJobState().toString());
-        } else {
-            Job job = jobManager.getJobById(jobId);
-            result.add(JOB_ID, jobId);
-            if(job != null){
-                result.add(JOB_STATUS, job.getJobState().toString());
-                String resultMessage = (String)job.getProperty("slingevent:resultMessage");
-                if(resultMessage != null){
-                    result.add("error", resultMessage);
-                }
-            } else {
-                // finished job
-                result.add(JOB_STATUS, Job.JobState.SUCCEEDED.toString());
-                JsonObject results = getJobResults(request.getResourceResolver(), jobId);
-                result.add(JOB_RESOURCES, results.getJsonArray(JOB_RESOURCES));
-            }
-        }
+        try {
 
+            String jobId = request.getParameter(JOB_ID);
+            if (jobId == null) {
+                Job job = submitJob(request);
+                result.add(JOB_ID, job.getId());
+                result.add(JOB_STATUS, job.getJobState().toString());
+            } else {
+                result.add(JOB_ID, jobId);
+                Job job = jobManager.getJobById(jobId);
+                if (job != null) {
+                    result.add(JOB_STATUS, job.getJobState().toString());
+                    String[] progressLog = (String[]) job.getProperty(Job.PROPERTY_JOB_PROGRESS_LOG);
+                    if (progressLog != null) {
+                        result.add(Job.PROPERTY_JOB_PROGRESS_LOG, String.join("\n", Arrays.asList(progressLog)));
+                    }
+                    String resultsPath = getJobResultsPath(job);
+                    JsonArray results = getJobResults(request, resultsPath);
+                    if(results != null){
+                        result.add(JOB_RESOURCES, results);
+                    }
+                } else {
+                    throw new ResourceNotFoundException(jobId, "Sling job was not found by id");
+                }
+            }
+        } catch(Exception e){
+            StringWriter sw = new StringWriter();
+            e.printStackTrace(new PrintWriter(sw));
+            result.add(JOB_STATUS, Job.JobState.ERROR.toString());
+            result.add(Job.PROPERTY_JOB_PROGRESS_LOG, sw.toString());
+        }
         try (JsonWriter out = Json.createWriter(response.getWriter())) {
             out.writeObject(result.build());
         }
@@ -145,26 +150,34 @@ public class ContentCatalogServlet extends SlingSafeMethodsServlet {
     }
 
     /**
-     * Read results of a completed job and returned parse it as JSON
+     * Read results of a completed job
      *
      */
-    JsonObject getJobResults(ResourceResolver resourceResolver, String jobId) throws IOException {
-        String resultsPath = getJobResultsPath(jobId);
+    JsonArray getJobResults(SlingHttpServletRequest request, String resultsPath) throws IOException {
+        ResourceResolver resourceResolver = request.getResourceResolver();
         Resource resultsNode = resourceResolver.getResource(resultsPath);
         if(resultsNode == null) {
-            throw new ResourceNotFoundException(resultsPath);
+            return null;
         }
         try(InputStream inputStream = resultsNode.adaptTo(InputStream.class);
             JsonReader reader = Json.createReader(new InputStreamReader(inputStream, StandardCharsets.UTF_8))
         ){
-            return reader.readObject();
+            JsonArray remoteItems = reader.readObject().getJsonArray(JOB_RESOURCES);
+            JsonArrayBuilder resources = Json.createArrayBuilder();
+            for(JsonValue val : remoteItems){
+                String path = val.asJsonObject().getString("path");
+                if(resourceResolver.getResource(path) != null){
+                    resources.add(val);
+                }
+            }
+            return resources.build();
         }
     }
 
     /**
-     * @return  the path to an nt:file resource with the job results as JSON
+     * @return  the path to a nt:file resource with the job results as JSON
      */
-    public static String getJobResultsPath(String jobId) {
-        return "/var/acs-commons/contentsync/jobs/" + jobId + "/results";
+    public static String getJobResultsPath(Job job) {
+        return String.format(JOB_RESULTS_BASE_PATH + "/%s/results", job.getId());
     }
 }