+++ bundle/src/test/java/com/adobe/acs/commons/contentsync/io/TestShardGenerator.java
@@ -17,5 +17,27 @@
  * limitations under the License.
  * #L%
  */
-@org.osgi.annotation.versioning.Version("6.15.0")
-package com.adobe.acs.commons.contentsync;
+package com.adobe.acs.commons.contentsync.io;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class TestShardGenerator {
+    @Test
+    public void testSharding(){
+        assertShard(3, "3");
+        assertShard(33, "33");
+        assertShard(333, "33/3");
+        assertShard(3333, "33/33");
+        assertShard(33333, "33/33/3");
+    }
+
+    void assertShard(int n, String expected){
+        ShardGenerator generator = new ShardGenerator("/var/test");
+        for(int i = 0; i < n; i++){
+            generator.nextShard();
+        }
+        assertEquals(expected, generator.getShard());
+    }
+}