+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ### Fixed
 
+- #3616 Content Sync: Fixed js errors when deleting configurations
 - #2745 Fixed circular dependency in EnsureOakIndex
 
 ## 6.14.0 - 2025-07-15
@@ -31,6 +32,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #3601 Content Sync: in case of an error print the exception and continue instead of aborting
 - #3596 Redirect Manager: com.adobe.acs.commons.redirects.servlets.* should expose error messages to end users
 - #3594 Redirect Manager: create parent structure if user enters a non-existing /conf path in Add Configuration.
+- #3626 Redirect Manager: Fix Replace Mode issue, delete only 'redirect-row' nodes
 
 ### Fixed