+++ CHANGELOG.md
@@ -20,7 +20,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ### Changed
 - #3623 Content Sync: improve error handling
-- #3638 Redirect Manager: fix search in tags property
+- #3627 Redirect Manager: Use bulk activation
+- #3638 Redirect Manager: Fix search in tags property
 
 ## 6.14.0 - 2025-07-15