+++ CHANGELOG.md
@@ -22,7 +22,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #3636 Content Sync: support dedicated egress IP address
 - #3631 Content Sync: refactor to use Sling Jobs
 - #3623 Content Sync: improve error handling
-- #3638 Redirect Manager: fix search in tags property
+- #3627 Redirect Manager: Use bulk activation
+- #3638 Redirect Manager: Fix search in tags property
 
 ## 6.14.0 - 2025-07-15